pyproject.toml
[project]
name = "llm-chess"
version = "0.1.0"
description = "The LLM Chess Tournament"
authors = [
    {name = "AutumnAurelium"},
]
dependencies = [
    "python-chess>=1.999",
    "openai>=1.45.1",
    "python-dotenv>=1.0.1",
    "requests>=2.32.3",
]
requires-python = ">=3.12"
readme = "README.md"
license = {text = "MIT"}

[build-system]
requires = ["pdm-backend"]
build-backend = "pdm.backend"


[tool.pdm]
distribution = true

[tool.pdm.scripts]
start = "python -m llm_chess"